Decouple WarnPunishedDelegate component
