+++ src/components/shared/warnPunishedDelegate/index.js
@@ -1,16 +1,46 @@
+import React, { useEffect } from 'react';
 import { compose } from 'redux';
-import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
+import moment from 'moment';
 import { getBlock } from '@api/block';
 import withData from '@utils/withData';
 import { withTranslation } from 'react-i18next';
+import VoteWarning from './voteWarning';
 import WarnPunishedDelegate from './warnPunishedDelegate';
 
-const mapStateToProps = (_, ownProps) => ({
-  isBanned: ownProps.isBanned,
-  readMore: ownProps.readMore,
-  pomHeights: ownProps.pomHeights,
-});
+const getPunishmentDetails = (punishedTimestamp, pomHeights) => {
+  const { start, end } = pomHeights && pomHeights[pomHeights.length - 1];
+  const startDate = new Date(punishedTimestamp * 1000);
+  const punishmentStartDate = moment(startDate).format('MM.DD.YYYY');
+  // 10: block slot interval, 60: minutes, 24: hours
+  const numOfBlockPerDay = 10 * 60 * 24;
+  const daysLeft = Math.ceil((end - start) / numOfBlockPerDay);
+
+  return { daysLeft, punishmentStartDate };
+};
+
+const Warning = ({ vote, ...props }) => {
+  useEffect(() => {
+    props.timestamp.loadData();
+  }, []);
+
+  const { daysLeft, punishmentStartDate } = getPunishmentDetails(
+    props.timestamp.data.timestamp,
+    props.pomHeights,
+  );
+
+  if (vote) {
+    return <VoteWarning {...props} daysLeft={daysLeft} />;
+  }
+
+  return (
+    <WarnPunishedDelegate
+      daysLeft={daysLeft}
+      punishmentStartDate={punishmentStartDate}
+      {...props}
+    />
+  );
+};
 
 const apis = {
   timestamp: {
@@ -24,7 +54,6 @@ const apis = {
 
 export default compose(
   withRouter,
-  connect(mapStateToProps),
   withData(apis),
   withTranslation(),
-)(WarnPunishedDelegate);
+)(Warning);
