+++ src/components/shared/warnPunishedDelegate/warnPunishedDelegate.css
@@ -34,7 +34,7 @@
   }
 }
 
-.editVoteContainer {
+.voteContainer {
   background-color: var(--color-warning-bg);
   border: solid 1px var(--color-warning-border);
   border-radius: 5px;
