+++ src/components/shared/warnPunishedDelegate/warnPunishedDelegate.js
@@ -1,5 +1,4 @@
 import React, { useEffect } from 'react';
-import moment from 'moment';
 import { withTranslation } from 'react-i18next';
 import PropTypes from 'prop-types';
 import { routes } from '@constants';
@@ -10,42 +9,15 @@ import { SecondaryButton } from '@toolbox/buttons';
 import Icon from '@toolbox/icon';
 import styles from './warnPunishedDelegate.css';
 
-const getPunishmentDetails = (punishedTimestamp, pomHeights) => {
-  const { start, end } = pomHeights && pomHeights[pomHeights.length - 1];
-  const startDate = new Date(punishedTimestamp * 1000);
-  const punishmentStartDate = moment(startDate).format('MM.DD.YYYY');
-  // 10: block slot interval, 60: minutes, 24: hours
-  const numOfBlockPerDay = 10 * 60 * 24;
-  const daysLeft = Math.ceil((end - start) / numOfBlockPerDay);
-
-  return { daysLeft, punishmentStartDate };
-};
-
-const getMessage = ({
-  isBanned, daysLeft, punishmentStartDate, t,
-}) => (isBanned
-  ? t('This delegate has been permanently banned from {{punishmentStartDate}}', { punishmentStartDate })
-  : t('Caution! This delegate has been punished for {{daysLeft}} days starting from {{punishmentStartDate}}', { daysLeft, punishmentStartDate })
-);
-
 const WarnPunishedDelegate = ({
   t,
   isBanned,
-  pomHeights,
-  timestamp,
   history,
-  editVote,
+  daysLeft,
+  punishmentStartDate,
   ...props
 }) => {
   const theme = useTheme();
-  const { daysLeft, punishmentStartDate } = getPunishmentDetails(
-    timestamp.data.timestamp,
-    pomHeights,
-  );
-
-  useEffect(() => {
-    timestamp.loadData();
-  }, []);
 
   useEffect(() => {
     if (history.location.pathname !== routes.account.path) {
@@ -53,24 +25,9 @@ const WarnPunishedDelegate = ({
     }
   }, [history.location.pathname]);
 
-  const message = getMessage({
-    isBanned, daysLeft, punishmentStartDate, t,
-  });
-
-  if (editVote) {
-    return (
-      <div className={styles.editVoteContainer}>
-        <Icon name="warningYellow" />
-        {t(
-          'Caution! You are about to vote for the punished delegate, this will result in your LSK tokens being locked for a period of {{lockedTime}} days. In addition, please note that your vote will not be counted until the {{daysLeft}} day period has expired.',
-          {
-            lockedTime: 'XX',
-            daysLeft,
-          },
-        )}
-      </div>
-    );
-  }
+  const message = isBanned
+    ? t('This delegate has been permanently banned from {{punishmentStartDate}}', { punishmentStartDate })
+    : t('Caution! This delegate has been punished for {{daysLeft}} days starting from {{punishmentStartDate}}', { daysLeft, punishmentStartDate });
 
   return (
     <FlashMessage
